+++ tests/NetXlsx.Tests/DisposedWorkbookMatrixTests.cs
@@ -28,6 +28,9 @@ public class DisposedWorkbookMatrixTests
         yield return new object[] { "Underlying", (Action<IWorkbook>)(wb => { var _ = wb.Underlying; }) };
         yield return new object[] { "AddNamedRange", (Action<IWorkbook>)(wb => wb.AddNamedRange("X", "S!$A$1")) };
         yield return new object[] { "NamedRanges", (Action<IWorkbook>)(wb => { var _ = wb.NamedRanges; }) };
+        yield return new object[] { "Protect", (Action<IWorkbook>)(wb => wb.Protect()) };
+        yield return new object[] { "Unprotect", (Action<IWorkbook>)(wb => wb.Unprotect()) };
+        yield return new object[] { "IsProtected", (Action<IWorkbook>)(wb => { var _ = wb.IsProtected; }) };
     }
 
     [Theory]